Added Exception

+++ video_downloader.py
@@ -3,6 +3,8 @@ import json
 import logging
 import argparse
 from pathlib import Path
+import glob
+import itertools
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s %(message)s')
 
@@ -12,25 +14,28 @@ def main(args):
     # headers = {'User-Agent': user_agent}
     #parts = ["train","dev", "test"]
     #for part in parts:
-    lan = "gse"
+    lan = "ase"
     
     #!cd /content/SP-10/dataset/gse
     r_path = Path(args.save_path)
     #Path("/content/SP-10/dataset/gse")
 
-    combined = []
-    for json_file in r_path.glob("*.json"): #Assuming that your files are json files
-        with open(json_file, "rb") as infile:
-            combined.append(json.load(infile))
-    combined_list = list(itertools.chain.from_iterable(combined))
-    with open(f"{r_path}/{lan}240.json", "w") as f:
-        json.dump(combined_list, f)
+    if Path(f"{r_path}/{lan}240.json").exists():
+      print(f"{lan}240.json exists.")
+    else:  
+      combined = []
+      for json_file in r_path.glob("*.json"): #Assuming that your files are json files
+          with open(json_file, "rb") as infile:
+              combined.append(json.load(infile))
+      combined_list = list(itertools.chain.from_iterable(combined))
+      with open(f"{r_path}/{lan}240.json", "w") as f:
+          json.dump(combined_list, f)
     
     
     logging.info(f"Downloading {lan} set")
     with open(f"{r_path}/{lan}240.json", "r") as f:
         data = json.load(f)
-    i=1
+    i=0, j=0, k=0
     for obj in data:
         #video_name = obj["video_name"]
         #sign_list = obj["sign_list"]
@@ -45,11 +50,17 @@ def main(args):
             logging.info(f"{file_path} already exists")
             continue
         file_path.parent.mkdir(parents=True, exist_ok=True)
-        logging.info(f"Downloading {file_path}")
-        logging.info(f"Requesting {video_url}")
-        urllib.request.urlretrieve(video_url, file_path)
-        i = i+1 
-    logging.info(f"Downloading {lan} set finished")
+        try:
+          logging.info(f"Downloading {file_path}")
+          logging.info(f"Requesting {video_url}")
+          urllib.request.urlretrieve(video_url, file_path)
+          i = i + 1  
+        except ValueError:
+          print("SKIPPED SKIPPED SKIPPED SKIPPED SKIPPED")
+          j = j + 1
+        
+        k = k + 1 
+    logging.info(f"Downloading {lan} set finished, {k} videos. {i} completed and {j} were skipped.")
